pull before push

+++ programming-principles-and-practice/24-numerics/christian/vortrag/prgs/limits/main.cpp
@@ -5,9 +5,9 @@ using namespace std;
 
 int main(){
 
-  cout << "DOUBLE_MAX:\t" << numeric_limits<double>::max() << endl
-       << "DOUBLE_MIN:\t" << numeric_limits<double>::min() << endl
-       << "DOUBLE_LOWEST:\t" << numeric_limits<double>::lowest() << endl; 
+  cout << "DOUBLE_MAX:\t" << numeric_limits<int>::max() << endl
+       << "DOUBLE_MIN:\t" << numeric_limits<int>::min() << endl
+       << "DOUBLE_LOWEST:\t" << numeric_limits<int>::lowest() << endl; 
   // check std::numeric_limits for more :)
   return 0;
 }
